pm: implement a new UI supporting manifests
The goal of the UI re-arrangement is to support manifests of the form already
used in the Fuchsia build system, specifically "dest=source\n".

Additional refactors are also done in this change to make way for changes in
upcoming pmd changesets, specifically moving much of the build process out of
the shared notion of pkg.Package.

Several UI bugs were also surfaced and manually tested. The most significant
changes to the formats output by the tools are:
 * meta/contents no longer contains any files in meta. The files were
   previously present in meta/contents because their names were not being
   passed to the signed message. That was a mistake in the signing process and
   was previously corrected. This will simplify the contents parsing
   requirements for Amber.
 * meta/contents now uses the same fundamental format as the manifests files.
   This format is insufficient for the ultimate term, but is fine for now. A
   discussion regarding "valid names" for tools such as the package manager is
   pending. It is likely the format will introduce either escaping or binary.
   The choice made herein is consistency until that decision is made.

Change-Id: I225b0330b4d06655cadb78dc97200ba4a047b64d