Delete unused SSDO acceleration code.
Keep the best variant and throw the others away; we can always spelunk
in the git history if we want to experiment with them again.  More
likely, when we revisit shadows we will consider completely different
approaches.

(note: there is still a compute-based SSDO sampler in ssdo_sampler.cc;
we spare it for now).

Change-Id: I6dfac91bfeb4dc814dea765eac81cda1c722f8f7