apps/media: fixes related to seeking past end of stream
SharedBufferSet was DCHECKing that it had buffers in LocatorFromPtr
before determining that the pointer in question was null (and
therefore didn't reference a buffer). Fixed that.

Also, the renderes weren't signalling the prime is complete when
they receive and end-of-stream packet. This occurs when we seek to
(or past) the end of the file. Given that they'll receive no more
packets, the renderers need to complete a pending Prime when they
see end-of-stream. Without this, the player gets wedged waiting for
the callbacks.

Change-Id: I6db26cdb41a4b6ba2c9c4df09f610a8ef5e4b29b