[kernel][arm64] correct the arch thread saved_exit_flags offset
This was silently corrupting whatever was at offset 24 in the arch section of the
thread structure, which was probably data_fault_resume. I believe that just so
happens to never actually be used at the same time as the saved_exit_flags field,
so it hasn't caused any trouble until now.

Change-Id: I221aaf59758408d093eb387283fc441109a6de1f