[ddk] new protocol support (1 of n)
This is the start of migration of device protocols from an vtable
structure to a structure with the vtable (ops) and context pointer
(ctx).  These will look like:

typedef struct {
    protocol_xyz_ops_t* ops;
    void* ctx;
} protocol_xyz_t;

The new device_get_protocol(...) fills out a structure of that
style.  The legacy device_op_get_protocol() works as before.
Once we've migrated all protocols, providers or protocols, and
consumers of protocols to the new style, we'll remove the legacy
interface.

The device hook get_protocol() is not used by any devices, so we
will convert it to the new style API now.

It's no longer possible to get the DEVICE protocol because that's
private between the devhost and the device and should not be
accessible to arbitrary clients.  Nothing depends on this.

This also converts MX_PROTOCOL_ETHERMAC and WLANMAC to the new
style and updates the DDKTL's wrappers of device, ethermac, etc.

Change-Id: I023c645312102066d301e605881055ee8b3086b7