[usb] Remove C99-isms from public usb interface.
Static array indicies in function arguments are a C99 extension which
is not available in C++ (at least, not as of C++14).  Our public
driver headers need to conform to the lowest common language
denominator WRT C/C++, so remove the use of this feature from the
ioctl_usb_set_interface IOCTL wrapper.

Instead, align this wrapper with the functionality which is already
exposed by the USB portion of the DDK.  Good news is that no one is
using this wrapper (as of today), so no further changes need to be
made.

Change-Id: I747b673693f0b135473bab007bac675d5b437609