Handle StartModule() being called again.
If StartModule() is called again with the same module name,
and the module url and link are the same as for the first call,
and no service exchange is requested, just connect the ModuleController
request to the existing module controller.

Otherwise tear down the existing module controller first and then
bring up a new one.

Exercised the new behavior in parent_child test.

Collateral cleanup:

* Make StartModule() and StartModuleInShell() methods more uniform.

Change-Id: Ib45f0ee09aebb6e14c48a473cb5277369214a7c3