[platform-device][bcm] introduce register_protocol()
Eventually we won't need platform_find_protocol(), because the new
open-related-device functionality will cover that.  But for now this
avoids the platform bus driver needing to touch internal mx_device_t
fields (especially the children list, which is unsafe for it to interact
with outside of the devhost lock).

No mechanism is provided to unregister protocols, as we'll be on the
new model before we consider unloading the mailbox driver.

Change-Id: I656b4a80469109077525839c563ba4bf287196cd