Importance changes on create and focus.
If a story is created or focused, trigger importance change
notifications, because importance changes then too.

If a story is deleted or defocused, no importance change is
notified, because the story is no longer shown anyway so its
importance should not matter (deleted), or nothing is logged
so the importance currently doesn't change (defocus). We may
change this in the future.

Change-Id: I756aa2096099e85710996029b2ce1d8267dccfcf