Set default headphone gain to 0.0dB
Now that there is a master gain control in the audio_server (which
defaults to -20 dB), we can move the default headphone gain for
Realtek codec headphone jacks from -30.0 dB back to 0.0 dB.  Most
users will be going through the mixer, and therefor will get the
default attenuation which is set there.  Developers who interact
directly with low level drivers during tests will need to be careful
to not accidentally hurt their ears.

Change-Id: I3393c239082ce832caecddef59dd88eea3ecd279