Provide API to modules to request story focus
Temporary hack. Add a RequestFocus() API to module context that allows
a module to request that their story and module take focus.

Current implementation explicity takes focus. We should be asking the
user shell, the current story shell, and module controller if it's ok
to interupt and switch focus to the background story.

Change-Id: I4cce25eb961908973e82450cbc6f66b8c8ccc8c0