apps/media: renderers unbind consumers when rebinding (MTWN-17)
Renderer's whose consumers were bound (via MediaRenderer.GetConsumer)
were DCHECKing when asked to bind again. This CL resets the consumer
(thereby unbinding it) if it's bound when GetConsumer is called. This
applies to both the audio and video renderer.

The old behavior was causing problems when rebuilding a graph,
because the creation of new connections was racing with the
destruction of the old bindings.

Change-Id: I968082ec46c30b60155c3fdd423ca43886e2fb7b