[audio_server] Shutdown outputs which fail sync write.
Shutdown an output which is using the older audio1 driver interface
and fails a synchronous write.  Right now, this is the best way we
have to figure out that an audio1 device has become hot-unplugged.
After the migration to audio2 is complete, we can use the channel
close event to figure this out.

Change-Id: Ia6b133905248f77627e5c642a459aa0d0212f3e4