[utest][threads] Add test for the case where a thread kills itself
This checks that mx_task_kill() does not return in this case.  Also
document that behaviour.  Previously, we didn't have any tests where
mx_task_kill() is called with a handle for the current thread.

This is a good way to test that the kernel's syscall implementation
checks THREAD_SIGNAL_KILL on the way out of the syscall.

Change-Id: If0d3c2eced82817de758d91a4520b2482929a715