[ddk] reorganize devhost and libdriver
Instead of libdriver being an indirection thing, it now is the
entirety of the devhost code.  This reduces overhead for the
various device_*() API calls, as they link directly to the
implementation on load now.

devhost (the executable) is simply main() that calls device_host_main()
provided by libdriver

We couldn't easily do this before, when devhost and devmgr were the
same program operating in different modes and some drivers were linked
into that directly, but in the new, tidier world, we can simplify things
a bunch.

Change-Id: I8ac8fe79ae03b7c5405143bc9b317a7199cac7e3