Reuse common page Operations in DeviceMapImpl.
This also resolves the issue that PageClients use the
wrong PageSnapshot. They can all always use the latest
snapshot, rather than the snapshot that accidentally
was the latest from the watcher at the time the operation
was scheduled.

Both Operations are now executed on the operation queue,
so Query() is now guaranteed to include the record written
in the constructor.

Change-Id: Iae7d23b50df70b606399f1b322a230c6d5af6cb0