Trying again: [context] Add partial_eq and filter_array filter options to ContextQuery.
This allows clients who wish to know, e.g., if a list of
'focal_entities' contains an object with "@type" == "some_type" to
express it in the query. They no longer must do this type of filtering
client-side.

This round of CL doesn't break wildcard queries. Added a test.

MW-96 #done

Change-Id: I01e73fa359719807dee578dfd0fe63f04c1abf2e