[devmgr] support adding new drivers at runtime
This is a placeholder while we sort out the longer time driver
management API.

Sending "add-driver:/path/to/driver.so" to /dev/misc/dmctl will
request that a driver be loaded.  If any unbound devices have
properties compatible with the driver's binding program, the
devmgr will attempt to bind it.

Change-Id: I36bfdd381bfdddbd7b5f08a08d53cc2b6a587df8