StoryController supports inspection of links and modules.
Add two method GetActiveLinks() and GetActiveModules() to
StoryController. Each returns a list of the respective items
and takes an interface pointer to a watcher for subsequent
updates to the returned list.

Add a method GetModuleController() to StoryController() in
order to be able to listen for module state changes on the
active modules listed by GetActiveModules().

Add code to test_user_shell that exercises the newly added
lists and watchers.

In order to make it simpler to obtain the module data from
the module context of active modules, pass a ModuleData
instance to the ModuleContextImpl constructor.

With this change, a StoryProvider client (specifically the
maxwell story_info agent) can keep track of all links
in a story and scan them for context links. Once we make
the changes there, we revisit the ergonomics of the current
API.

FW-219

TBR=vardhan,alhaad

Change-Id: I03509ce786df447b74bb890a85d1744351d61031