Store user's display name, profile url and image url to users.db.
When adding id_token for a newly added user we also get the user's
display name, profile url and image url. We store it in users.db and
give it to user shell.

This also means that we start using the display name fetched by token
manager and not the one provided by device shell.

Change-Id: I0fc3d3859cc633e2264850b8629fe9a63af8d265