Do not reinflate externally-added modules which have already finished.
This is done by adding a 'stopped' state in the legder-persisted
ModuleData struct, and checking for it to filter out externally-started
modules.

This fix is still incomplete: we need to also capture
ModuleController.Stop() as a reason to persist the 'stopped' state, and
probably also for internally-added modules.

FW-210 #done

Change-Id: I75c0860c45bccb44e21972862017625fc7b3e5c0