[ulib][musl][mxio] __sanitizer_get_configuration via loader service
This adds a <magenta/sanitizer.h> interface for getting a read-only
VMO for a debugging configuration file.  The libc implementation of
this function works by simply sending the configuration name in an
RPC to the loader service.  Currently this is treated on the server
side just like script interepreters, i.e. just read in from an
absolute file name.

Change-Id: I08aa5aa42ebc376a510ae9d5d7b69162a47ec037