[udev][usb-xhci] Fix iotxn leak that could occur when handling errors
The XHCI controller can return errors for any TRB in the transfer ring.
When that occurs, we may have to search further down the ring to find the
"event data" TRB which contains a pointer to the iotxn for the failed transfer.
We used to only search 5 TRBs, which is now insufficient given that
scatter/gather support can split a transaction across many TRBs.
So now we search the full length of the transfer ring for the "event data" TRB
to ensure that we can complete the iotxn for the failed transfer.

Change-Id: I286d6903b6858e57c84c296763e371e6c6e8a440