apps/media: add lib/audio/perceived_level.*
This CL adds a helper class that provides translation between gain
values and 'level' values suitable for volume sliders. A float
level in the range [0,1] is supported as well as an int level in
the range [0,max] where the caller specifies max. Levels are
logarithmic, like gain, but the range is adjusted to meet user
expectations. The extremes of the level ranges are explicitly
mapped to -160db (effective silence) and 0db.

Change-Id: I01c668e6e15aec1f996f4358e99a516b433efdc6