Add support for default values in filters
This is a solution for the problem that XdrRead() DCHECKs if a
new field is added to a filter. When this happens within the
Framework, the process dies and the user is returned to the
login screen with no visible explanation. Last time this happened
we told everyone to delete all of their stories, which is a
significant time sink for people to discover what broke and
how to fix it.

Mesch and I concluded that a DCHECK is still the best behavior to make
sure that fields aren't initialized improperly, but we didn't have a
solution for the case where there's a reasonable default.

This CL adds the ability to create defaults for fields in filters.
This is done by initializing the field in the data structure, then
ignoring any error that might happen in the Field() function.

I considered adding variants of the Field() function that take a
default value, but a "proper" solution requires more design work
and this solution localizes the changes so they are easy to
modify later.

Change-Id: Iec3f2abe395e393752cfba67a9eb404dc448d3ba