[kernel] Fix a race in reporting process exit statuses
Before this change, it was possible for a process exit status (as
reported by mx_object_get_info()) to change.

Fix this by not changing the exit status after it is first set.  This
changes ProcessDispatcher::Exit() to use the same logic as in Kill().

Add a test which reproduced the issue with:
 * ProcessDispatcher::Exit() changed to call thread_yield()
 * sched_get_top_thread() changed to select a thread at random

Change-Id: I14f5da462fa1feb3892725500983b9919a93f574