Fix XDR filter error handling to only process read errors
In my first pass of default values, the filter wouldn't know
whether the current op was read or write and so overwrote user
values when trying to supply a default. In this updated version,
a callback is provided to set the default value if an error
happens while reading the JSON field.

Change-Id: I30d6a98a1a835fd8c55c6ab1ac36d019b60b4fc7