Record ModuleContext.Stop() module state in the ledger.
Change ModuleControllerImpl::Stop() to delegate to
StoryControllerImpl::StopModule(module_path) rather than to call
ModuleController::Teardown() directly.

Implement StopModule(module_path) as an Operation on the operation
queue of the story controller.

This StopModule operation also writes the stopped bit into the module
state, moving the machinery from ModuleWatcherImpl and eventually
calls Teardown() on the module controller impl.

ModuleWatcherImpl calls StopModule() too.

The StopCall of the story controller continues to call
ModuleControllerImpl::Teardown() directly and thus bypasses
the writing of the stopped bit in the module state.

One test needed an increased timeout, because the StopModule
procedure involves a Defocus call to the story shell, which is
very slow if it happens shortly after startup and story shell
is not yet running. Previoulsy this slow call was just bypassed
by the concurrently executing StopCall of the story controller.
Now, however, StopModule is an Operation too and StopCall waits
for it to finish.

Collateral cleanup:

* Rename TakeOwnership() to AddModuleWatcher().

* Add variant of StoryStorageImpl::WriteModuleData() that takes
  on ModuleData at once.

* Add missing Ready() call to one test module.

FW-210 #done
FW-219

TBR=vardhan,alhaad

Change-Id: Ie9c53874a67d8f56fe4432e4a9a89e4491f8ca65