Generate constexpr constants in C++ bindings.
Declare constexpr constants (instead of simply const) for all numerical data
types when generating C++ constants.

In theory, this can be done for other POD-like types (types like a
std::chrono:duration, for example), but figuring out when that is safe or not
can be tricky and is currently left as an exercise for the reader.

Change-Id: I4b6bf41c72e53cb79a9ad85e994b68e4edd7e91e