Disable checks for Guest accounts.
OAuth Token Manager service assumes account to be a valid Gaia account
for retrieving tokens. But SysUI has a Guest mode that generates an
account which has an Id but no other indication that it is a Guest
account. For such accounts, return success with an empty response for
now.

In the future, such accounts must be marked as Guest before talking to
OAuth Token service.

Also replaced FTL_LOG(INFO) with FTL_VLOG(1) to reduce log spam.

Change-Id: I914a64a01d69f114788536fb5902e27adb41e916