[syscalls][socket] Update docs and test for mx_socket_write() error conditions
Fix the docs: (MX_)ERR_TOO_BIG does not exist as an error code.

Furthermore, mx_socket_write() doesn't seem to have a code path for
the exact condition of "*size* is larger than the largest allowable
size for socket messages".

The actual error behaviour is already covered in the docs under
"DESCRIPTION".

Change the test to be stricter: Test the status code, and test whether
'written' was or wasn't filled out.

Change-Id: I257695477ac673b24cb108728f96a2e207942733