apps/media: make video progress immediately after seek (MTWN-18)
This CL fixes an issue in which, after a seek, video sometimes takes
awhile to start progressing. This happens because the seek position
is well before an i-frame, and the decoder has to wait for an i-frame
before it knows how to start decoding.

The fix is to seek to the last i-frame that occurs at or before the
seek position and to have the renderers discard all the packets up to
the seek position. This allows the video decoder to have the i-frame
it needs, but it also means that the graph has to fast-forward from
the i-frame to the seek position.

The ffmpeg demux supports a flag on its seek function that finds the
correct i-frame.

TimelineControlPointImpl, which manages timeline changes and priming,
has the seek position it needs when priming is requested. This is a
bit of a coincidence, design-wise, but it's reliable. The control
point now provides that seek position to the renderers when prime is
requested, and the renderers throw away packets until the seek
position is reached.

This approach is more or less consistent with the new flow
management design due to be implemented soon, but the new design
will be more explicit about how this is done. The signal equivalent
to a prime request in the new design actually contains the desired
seek position explicitly. Another difference is that unneeded packets
will be discarded much further upstream in the graph, making seeks
more efficient.

Change-Id: I845da96767f38574ec1f34a6e17b2c0ea5be1795