[musl][socket] Remove the MSG_NOSIGNAL flag
The MSG_NOSIGNAL flag is intended to suppress delivery of a SIGPIPE
signal on writing into a socket that isn't connected. Since we don't
have signals, this is not needed.

MSG_NOSIGNAL was only very recently added to POSIX, and several
platforms don't have it. As such many systems already have a mechanism
for checking for the existence of this flag and changing build
behavior accordingly.

Change-Id: I84829646d971c3f8b8d8edb933845cabd8d13a10