Give a diagnosis hint on socket.error
When run_tests establishes a connection to Fuchsia, and then gets a
socket error, it's often the case that Fuchsia has failed in some way.
This is especially true on the bots, where link-local networking in QEMU
has been reliable.

I've been getting bug reports where people assume this is an
infrastructure problem, because it says socket error, which sounds like
infrastructure. So we need a hint to look at the log.

Tested this change by killing QEMU in the middle of a test run.

IN-64

Change-Id: I7f08cbd91f45cc582f344c576f46ca2e9c84abb9