[ulib][musl][ldso] Remove ldd_mode and associated special cases
The ldd_mode support doesn't actually work and we don't really need it.
Simplify code by removing it.  This also removes the special-case kludge
for known library names.  We don't need this for lookups of "libc.so" to
work, and we don't care to support lookups of "libpthread.so" et al since
we don't actually have any DSOs by those names and aren't trying to provide
loose compatibility with dlopen("libpthread.so", ...) and the like.

This exposed a bug in the logic for lookups of the "detached" libraries,
i.e. libc itself and the vDSO.  That's fixed by the introduction of the
detached_head variable.

Change-Id: Iab51ebaa073dba24cf671987c6d7e88205eee9b2