[ulib][musl][asan] Don't read off the end of strings under ASan
The string functions written in C are optimized to do aligned word
accesses, even when this might read off the end of the string (by up
to one byte less than the word size, i.e. 7 bytes).  This is fine in
general, since alignment means that this read can never cross a page
boundary.  However, ASan detects any access off the end of an object
by even a single byte.

Change-Id: Iffe813d43b41b6cf94824614260584c84efce1de