Test for context links.
Context links are implemented in
https://fuchsia-review.googlesource.com/#/c/36165/.

Test runs as user shell. The user shell starts a module that posts
a @context link, and the user shell registers a context listener
that picks up the resulting context topic.

TODO in followup: If the context listener is registered too early,
the context listener connection closes and no context updates are
received. Laster registration works; marked as HACK().

FW-219

Change-Id: I183e469fa8d5b6368486327aeb39d316726e07e2