apps/media: support gain controls with no distinct mute
This CL changes the behavior of the audio policy service to better
support system gain UI that doesn't have a separate mute control.
It also improves documentation for related methods and tools.

The audio policy service exposes system audio gain and system audio
mute separately. This can be useful when the user wants to mute
audio but doesn't want to change the gain setting that will be
restored when she unmutes.

On the other hand, designers may opt for a simpler gain control that
doesn't have a separate mute control. In this case, the user must
turn gain all the way down to mute audio.

Both of these approaches are now supported by the audio policy
service by virtue of these behavior changes:

1) When gain is set to -160db from a higher value, mute is set to
   true implicitly.
2) When gain is set to a higher value from -160db, mute is set to
   false implicitly.
3) When gain is -160db, attempting to set mute to false has no
   effect (mute remains true).

In the near future, docs will be added that guide developers to the
services, libs and tools relating to system audio gain.

Change-Id: Ifc1e39e24cf215b8e52bde355ef16dfd4bb63c4c