[ulib][musl][mxio] __sanitizer_publish_data via loader service
This adds a <magenta/sanitizer.h> interface for publishing a VMO
to a "data sink".  This will be used for exporting binary blobs
such as coverage data.  The libc implementation of this function
works by simply sending the VMO and data sink name in an RPC to
the loader service.  Later this will be implemented on the server
side to do something with the data.

Change-Id: I21a25fe66445a0ef5e26bcd3da9e73d1a0ba448d