[fs] Add support for O_ADMIN, restricting unmounting capability
O_ADMIN is a new flag which is provided to connections to
new filesystems. This flag indicates that the client which
has the iostate connection has the capability to demand that
the filesystem should shut down. Previously, ANY open connection
to a filesystem (file, directory, etc) would be sufficient to
terminate the entire filesystem. With the addition of the
O_ADMIN flag, however, unmounting capability is restricted to
the connection that is typically pinned to the mounting vnode.

Test added in fs-management, blobstore tests updated (connections
to single blobs can no longer unmount the blobstore).

Change-Id: Ib371f10e6ddd446ebc25ec513a3d7c1863eca30a