[dart] Convert labels to package names at build time to speed up generation
Creating a subprocess during GN generation is expensive, especially on
OS X, and creating one for each of our many Dart packages adds up to
a substantial cost on the generation step. Since the purpose of
computing Dart labels is to pass them to scripts that run at build
time, this moves the computation into these scripts where the cost of
creating the process is already paid.  This speeds up generation of a
default project set from 6700ms to 5193ms on a MacBook Pro. Excluding
the time spent running the 'gn check' operation, and excluding rust
generation (which is excessively expensive for similar reasons), this
speeds up generation time from 4571ms to 2819ms, which is dominated
by Dart VM build rules.

Change-Id: I297713f3b3bdbba6ba25751a8f8f539e6767c15e