Suppress story state notifications during story delete.
Because the story record is deleted before the story is actually stopped,
story state changes from stop for delete may reach clients after they have
seen the delete call. We still cannot simply reverse the order of story
stop for delete and delete story info from ledger, because we still need
to stop the story for delete when we receive the delete from the ledger
page on other devices (this might change as we learn more about cross
device operation, but it's how it is currently).

Therefore, story state notifications are suppressed altogether during
DeleteCall execution. They are also suppressed during StopForTeardown(),
as nobody is expected to listen for them there anyway.

Added a test case to the user shell test, and verified that the user shell
test fails without this fix.

Collateral cleanup:

* Flagged one more race condition as TODO: When a StopForDelete() is
  in the operation queue front of a StopForTeardown(), the latter
  will not complete.

* Flagged as a TODO that we actually need to delete the story page on
  DeleteStory().

* Some comments brushed up.

FW-222 #done
FW-217

Change-Id: Ibd4aa46f59d4295ee1df54e5d42aa69813196364