[ddktl] Preserve null-ness when fetching name()
If a user of the ddktl Device<> helper class is in an unregistered
state (eg. Add has not been called), do not call through to
device_get_name during the name() accessor.  Right now, the DDK's
implementation of device_get_name will blindly access the name member
of the opaque mx_device_t struct.  This means that passing nullptr for
the device will result in a pointer coming back which is non-null, but
also not valid.  Instead of allowing this, simply return nullptr if
the device is null during name().

Change-Id: I953812f8b12c6cced7222120e7d53f6dfd4c0b3b