Increase audio2 command timeout.
Some devices can take a significant amount of time to respond to some
commands which get issued.  In particular, I've run across a set of
B&W USB speakers recently which take ~104mSec to power up their
amplifier when switching to their alternate audio interface when
starting DMA.  The command timeout at the mixer level was only
100mSec, causing it to give up before the speakers could fully turn
on.

For now, increase the timeout to 500mSec.  As time goes on, the
magenta output code is going to transition away from pseudo
synchronous mx_call ways of interacting with the audio drivers and to
a fully asynchronous state machine design.  We can re-evaluate what
reasonable timeouts are for various operations at that point in time.

Change-Id: Idc83b62190bddada42a01c00a883623d84c30069