Use right LinkWatcher semantics.
We got LinkWatcher semantics wrong in
https://fuchsia-review.googlesource.com/#/c/36165/.

A LinkWatcher connection is dropped when the Link connection it
was registered at is dropped. Therefore, we cannot drop the Link
connection after we register the LinkWatcher. We did this before
because we did not want to keep the Link in memory beyond usage,
but for now we have to.

This is another argument why LinkWatcher should *not* be used
by clients other than modules, which would solve this dilemma.

Another possible solution would be to introduce another kind of
watcher, which is not tied to its Link connection. A watcher
registered with WatchAll() could have this behavior. TBD.

A test for this fix is in
https://fuchsia-review.googlesource.com/#/c/36467/.

FW-219

Change-Id: Icf00cdbc02fdf048a115b688488549392fd599b5