[dev][usb][uapp][hidtouch][ulib[hypervisor] Clean up use of C99 inline
The rules of thumb for C are:
    In a header file, use 'static inline'.
    In a C source file, just use 'static' and the let compiler optimize.
Using 'inline' without 'static' has subtle semantics that are not what
anyone usually is trying to do, and differ from the C++ semantics.

The bare 'inline' uses in these C headers and source files meant that
things can fail to link if the compiler makes different optimization
decisions than you were guessing it would.

Change-Id: Ie923e785cfda47d32d49cc41d736e2ec34f1d91f