Make LinkWatcher more flexible.
A LinkWatcher connection now survives the Link connection it was registered on.

To accomplish this, the LinkWatcher is now kept in LinkImpl, together with a
weak ptr of the LinkConnection it was registered on, rather than in the
LinkConnection.

Explicitly mention this in the Link fidl documentation.

Collateral cleanup:

* Brush up all fidl documentation comments of Link.

FW-219

Change-Id: Ib0553a6e3452aa03b23da16687aa24d88735eba9