[kernel][rng] Add entropy collector test
Add a special build of Magenta that will pause during boot to test
entropy collectors. It gathers approximately 1 Mbit of random data into
a buffer, then proceeds with boot. Later the data will be copied to a
VMO and exposed in the filesystem as /boot/kernel/debug/entropy.bin.

To build the entropy collector test system, set the
ENABLE_ENTROPY_COLLECTOR_TEST preprocessor flag, either by adding
    EXTERNAL_DEFINES := ENABLE_ENTROPY_COLLECTOR_TEST=1
to local.mk, or by adding
    EXTERNAL_DEFINES=ENABLE_ENTROPY_COLLECTOR_TEST=1
to the call to make.

When ENABLE_ENTROPY_COLLECTOR_TEST is not set, no extra kernel memory is
allocated and no debug file is produced.

Change-Id: I55e6d5bb6323b5434e407b12ca14ddc79fe9237d