[mxtl] Fix destructor ordering in the SlabAllocator<>
Roland found this one while standing up ASAN.  The SlabAllocator
desctructor was going over its list of slabs and freeing before then
sanity checking the free object list, and then clearing out free list.

Clearing out the free list is a "clear_unsafe" operation, meaning that
it just nulls out the head pointer of the free list (which prevents an
improper free operation when the list is made of managed pointers and
the list goes out of scope), so it didn't actually touch any of the
objects which were freed as the slabs were destroyed.

The sanity check (in an MX_DEBUG_ASSERT) OTOH was comparing the free
list size to the number of objects which had ever been allocated.  As
our intrusive lists do not hold their size internally, this involved
an O(N) size_slow() operation which walks the list to determine its
length.  This operation was touching memory which had already been
freed when the slabs themselves were freed.

Avoid this by...
1) Computing the free list size (in a debug build) first and stashing
   it.
2) Clean up the free list.
3) Free the allocated slabs, computing the number of objects which had
   ever been carved out of slabs in the process (again, debug build)
4) Finally, assert that the size of the free list at the start of
   destruction was equal to the number of objects we had ever carved
   out of slabs over the lifetime of the allocator.

Change-Id: I87fab80ba8ff68450ac6179c8dc0d4d6cde8ae91