[ports-v2][exceptions] reduce size of mx_exception_packet_t
To enable migrating exception delivery to ports v2, we need exception
packets to fit in the 32 byte limit of ports v2 payloads.  This change
reduces the exception packet to the minimal identifying information
(type, pid, and tid), with 16 bytes of expansion space for future
features.

Userspace callers are updated to adjust to the new packet format and
to obtain the mx_exception_report_t from mx_object_get_info() where
it is needed, instead of depending on it being inline in the packet.

Change-Id: I3018917f97cadb0f8b673c24dd96fa652096e583