[minfs] Use initialized declarations for bcache
In the case that MinFS is mounted on a block device
which already has a FIFO-based client, the block device
will reject the "new" request to "ioctl_block_get_fifos".
When this happens, the Bcache destructor is executed,
which conditionally calls "block_fifo_release_client"
if "fifo_client_" is nonnull.

This "nonnull" check is intended to answer the question,
"has the fifo_client_ been allocated?" (in the case above,
it has NOT been allocated). Unfortunately, without
explicit null initialization, the fifo_client_ may or
may not appear to be allocated, and may actually be
"freed" when it is an uninitialized variable.

Change-Id: Ia1008e91c41b992af2b379cb49c624dc1f85577f