[boringssl] Add secondary/third_party/boringssl
This CL moves BUILD_template.gn and generate.py from
//third_party/boringssl/fuchsia to here.  This will facilitate reverting
//third_party/boringssl to a pure mirror of upstream.

It's noted that the approach taken with generate.py could stand a lot of
improvement, but for now the only modifications to that file are to make
it independent of its location.  Further enhancements will be handled in
later CLs.

Change-Id: I9ead7d45294d9fbafb964e7b59beccc707829b3b