apps/media: introduce TimelineControlPoint.SetProgramRange
A previous CL (https://fuchsia-review.googlesource.com/#/c/35899)
changed seeking behavior to make the demux seek to the i-frame
before the requested position. The graph then fast-forwards to
the requested position, the renderers discarding packets that
precede the requested position.

The renderers' fast-fowarding behavior requires the renderes to
know the requested position when they prime. CL 35899 depended
on a coincidence - the renderers get a SetTimelineTransform call
that contains the requested position before they get a Prime call.

Unfortunately, this isn't always the case. When switch from one
source to another (playing new content), there is no
SetTimelineTransform call before the Prime call that would provide
the new requested position (0). The effect was that, after switching
sources, the renderers would throw away content until the new source
caught up to the position at which the old source stopped. From the
user's perspective, the progress bar advances during that interval,
but no audio or video are renderered.

This CL addresses the problem by introducing a new method exposed by
renderers (in their capacity as TimelineControlPoints). That method
is called SetProgramRange. It takes a program id and a min and max
PTS. The program id is unused for now - it is used to distinguish
distinct content which, in the future, will be able to flow
consecutively through the graph. For now, the program id will always
be zero. We also don't currently need the max PTS, which will always
be kMaxTime.

SetProgramRange is sent prior to Prime to let the renderers know what
packets to discard to get the correct fast-forwarding behavior. In
particular, SetProgramRange is called with a min PTS of 0 after
changing sources and before Prime is called.

A subsequent CL will add FLOG logging for this method.

Change-Id: Ib6b8e1ec376dbcb7e2970ae45e40f9063f21e1e8