[unittest] Initialise the all_ok field as early as possible
Suppose we have a test case like this:

  bool test_foo() {
      BEGIN_TEST;
      ...
      END_TEST;
  }

Before this change, all_ok is initialised in BEGIN_TEST.

After this change, all_ok is initialised immediately before test_foo()
is entered.  This reduces the risk that all_ok is accidentally used
uninitialised, e.g. by omitting BEGIN_TEST.

This change reduces the amount of work that is done by macros.

Change-Id: I912eb278900ed9a6a5e9568fd5016bdd34018ad0