[boring-crypto] Match boring-crypto to boringssl
This CL aims to match ulib/boring-crypto's directory structure to
boringssl's, as well as reducing the number of modifications to
boringssl's code to a minimum.  Both of these changes reduce the cost of
maintaining a partial clone of boringssl, making it easier to add
additional algorithms in a maintainable manner.

It also separates out the kernel unit tests into lib/boring-crypto.
These tests are based on boringssl's tests, but have been rewritten to
integrate with Magenta.

Finally, it adds a script to compare Magenta's boring-crypto with the
third party BoringSSL.  It is aware of the allowable modifications and
will report all other files that do not match.  It should be used when
rolling BoringSSL to check if any changes needed to be ported to
Magenta.

Change-Id: Iecfefad5c929f321ac02a960df9d74bd86c3dbc9