[boring-crypto] Enable assembly for ChaCha20
This CL pulls in the assembly implementation for ChaCha20, and the
associated machinery needed to make it work.  In particular, it
provides a stub implementation in cpu-aarch64-magenta.cpp.  All
other files under crypto/ and include/ match those in BoringSSL.

It also add scripts/perlasm.sh, which can be used to regenerate the
assembly code from BoringSSL's perl scripts when needed on updating.

Change-Id: I37c7f38d614828bdd75fae44fb10f5e828952933