[unittest] Make "msg" argument optional for ASSERT/EXPECT_EQ etc. in C++
Before this change, ASSERT_EQ() and EXPECT_EQ() require a message
argument, but this is often not very useful, and a lot of tests pass
an empty string, e.g.

  ASSERT_EQ(Foo(), MX_OK, "");

rather than, say:

  ASSERT_EQ(Foo(), MX_OK, "Foo() failed");

Requiring this argument adds visual noise to the code, and it can slow
down writing tests if you often forget to add "".

This change makes the argument optional, at least in C++, so that we
can just write:

  ASSERT_EQ(Foo(), MX_OK);

This is implemented using variadic macros and C++ function
overloading.  We could implement this for C, too, using a var-args
function, but I hesitate to do that because it would be less type
safe.

Change-Id: I05d671fe15260a8eaeb2fdf925f40d58be2e199a