[resources] massively simplify resources
Resources now are simple access tokens with a 32bit "kind" and
a 64bit "range".  If you have the WRITE right, you can create
a child resource, provided that child is of the same kind and
of a range that fits within the parent's range.

The kernel gives userspace a single "root" resource which may
be used to create any kind of resource.

Change-Id: I4817b7b787386059e02a36ee235723bfb1244934