[tools][netaddr] Handle multiple network interfaces
If a single device is connected to a host with multiple active network
interfaces, netaddr incorrectly claims there are multiple devices.
Since it is the same device with the same node name, specifying the
node name is not a successful workaround. This happens in particular on
a Mac connected to the device over both WiFi and Ethernet.

This change uses the node name as an easy check to see if the same
device is being found multiple times.

Change-Id: Ibda8b1739b150bcf406301c92940b1afb1f67ecf