[tools][mdi] Add support for integer expressions and constants
MDI now supports the arithmetic operators '+', '-', '*', '/', '%'
as well as the bitwise operators '~', '|', '&', '^', '<<' and '>>'.
The syntax and precedence rules are the same as in C.

Integer constants can now be defined for integer literals or expressions.
For example:

const ONE = 1
const THREE = 1 + 2

Due to the ambiguity introduced by the '-' operator, commas are now needed
for separating array elements.

MDI test updated to include expressions and constants,
as well as tests for zero length lists and arrays.

Change-Id: Iaa3b020cc8d4d79738c3eacedd9f58f11e8184e1