[syscall] Work-around recursive locking in vmo_read/write
There is a known deadlock in the VmObjectPaged::ReadUser/WriteUser, that
has not yet been resolved (MG-730).  The deadlock is a cyclic locking
issue that occurs when page faulting while performing a usercopy in the
middle of a ReadUser/WriteUser.  This results in a panic if the second
lock involved is the same as the first.

Thanks to Tim Newsham, NCC Group for reporting this.

MG-867 #done
MG-730 #comment Add workaround for process_read/write_memory

Change-Id: I28710a67ac26e4d4c9e35ac8b39237549b2ee335