Revamp gpu mapping cache
The working set of graphics memory for fuchsia turns out to
be very very large, so the lru scheme leads to cache thrash
without blowing up the memory cap, which could lead to us
potentially retaining references to a lot of physical memory.

Instead we opt for a simple scheme of caching the mappings for all
buffers on a given connection while the connection has that buffer
imported.

MA-288 #done

Change-Id: Ib0f55e5008f436c3bcef32392e4052662b243e60