[mxio] Fix open path adjustment bug for absolute paths
The following commands should all display the same information,
when executed from the root directory:
  - 'ls data'
  - 'ls /data'
  - 'ls //data'

Unfortunately, prior to this patch, they did not.

__mxio_open_at and __mxio_opendir_containing_at open paths
in two phases. First, the incoming path's 'base' iodir is determined:
one of either CWD or the root directory for the calling process.
Afterwards, the path is canonicalized on the client side before
being transmitted to the remote server.

The "mxio_iodir" function, which determines the path's 'base',
adjusts the incoming path if it is absolute, in preparation
for the request to be sent to the server. However, if
the calling process requests "//data", then that request is transformed
into "/data", which still won't exist on a server (it would
be "data", with no starting "/").

To fix this issue, skip ALL the preceding "/" characters
the appear in the path prefix, rather than just one.

Change-Id: I86ae8662776b43af7d0eb884ddf443380251c5ef