[ddk][usb] Remove USB driver support code from the DDK library
Most of the functionality in ddk/common/usb.c has been moved to the USB protocol.
The remainder (usb_iotxn_alloc() and the USB descriptor iterator support) is now
in libdriver.

Function names remain the same, but now take a usb_protocol_t* instead of a
mx_device_t* as its primary argument.

Change-Id: I36d4074eeea5b0c53028636a73dd3910476415af