[kernel] Remove no-longer-needed racy status check in ProcessDispatcher
The re-entrant case described (in the comment being removed here)
should no longer occur, because ProcessDispatcher::SetStateLocked() no
longer closes the process's handles while holding state_lock_.
Instead, it uses ReapHandles(), which closes the handles on another
thread.

The use of state_ in on_zero_handles() is technically racy because
this is done without holding state_lock_, and state_ can be changed by
another thread.  So it's good to remove this, and remove a use of
TA_NO_THREAD_SAFETY_ANALYSIS.

Change-Id: If4abb795f89756e6327f754792f601c412c7e977